refactor(cli): extract openclaw config controller
